Added: KeyPressedListener, a stripped KeyEventListener to only listen for KeyPressed. Kund, a class with static name, for now, of logged in user.
Modified:
LoginPanel: Implemented and commanding MainFrame to switch to UI_ADDToCart on successfull Login and stores username in static Kund.
MainFrame: method for changing panel in view. and made it singleton.
Repository: Added catch-clauses and removed throws.
Skobutik: now used getInstance of MainFrame because of singleton.
UI_AddToCart: removed a lot of catches and throws. Also added a Commented ActionListener for when it's time later, MouseEvent is easier for testing.